moved FeatureApplicationError to new package
