+++ src/compas_timber/_fabrication/double_cut.py
@@ -15,7 +15,7 @@ from compas.geometry import intersection_plane_plane_plane
 from compas.geometry import is_point_behind_plane
 from compas.tolerance import TOL
 
-from compas_timber.elements import FeatureApplicationError
+from compas_timber.errors import FeatureApplicationError
 
 from .btlx_process import BTLxProcess
 from .btlx_process import BTLxProcessParams
@@ -296,7 +296,7 @@ class DoubleCut(BTLxProcess):
 
         Raises
         ------
-        :class:`~compas_timber.elements.FeatureApplicationError`
+        :class:`~compas_timber.error.FeatureApplicationError`
             If the cutting plane does not intersect with beam geometry.
 
         Returns
