+++ src/compas_timber/_fabrication/dovetail_mortise.py
@@ -13,7 +13,7 @@ from compas.geometry import intersection_line_plane
 from compas.geometry import is_point_behind_plane
 from compas.tolerance import TOL
 
-from compas_timber.elements import FeatureApplicationError
+from compas_timber.errors import FeatureApplicationError
 
 from .btlx_process import BTLxProcess
 from .btlx_process import BTLxProcessParams
@@ -502,7 +502,7 @@ class DovetailMortise(BTLxProcess):
 
         Raises
         ------
-        :class:`~compas_timber.elements.FeatureApplicationError`
+        :class:`~compas_timber.errors.FeatureApplicationError`
             If the cutting planes do not create a volume that itersects with beam geometry or any step fails.
 
         Returns
