+++ src/compas_timber/_fabrication/dovetail_tenon.py
@@ -14,7 +14,7 @@ from compas.geometry import intersection_line_plane
 from compas.geometry import is_point_behind_plane
 from compas.tolerance import TOL
 
-from compas_timber.elements import FeatureApplicationError
+from compas_timber.errors import FeatureApplicationError
 
 from .btlx_process import BTLxProcess
 from .btlx_process import BTLxProcessParams
@@ -607,7 +607,7 @@ class DovetailTenon(BTLxProcess):
 
         Raises
         ------
-        :class:`~compas_timber.elements.FeatureApplicationError`
+        :class:`~compas_timber.errors.FeatureApplicationError`
             If the cutting frames do not create a volume that itersects with beam geometry or any step fails.
 
         Returns
