+++ src/compas_timber/_fabrication/drilling.py
@@ -17,7 +17,7 @@ from compas.geometry import is_point_in_polyhedron
 from compas.geometry import project_point_plane
 from compas.tolerance import TOL
 
-from compas_timber.elements import FeatureApplicationError
+from compas_timber.errors import FeatureApplicationError
 
 from .btlx_process import BTLxProcess
 from .btlx_process import BTLxProcessParams
@@ -292,7 +292,7 @@ class Drilling(BTLxProcess):
 
         Raises
         ------
-        :class:`compas_timber.elements.FeatureApplicationError`
+        :class:`compas_timber.errors.FeatureApplicationError`
             If the cutting plane does not intersect with the beam geometry.
 
         Returns
