+++ src/compas_timber/_fabrication/step_joint_notch.py
@@ -14,7 +14,7 @@ from compas.geometry import intersection_plane_plane
 from compas.geometry import is_point_behind_plane
 from compas.tolerance import TOL
 
-from compas_timber.elements import FeatureApplicationError
+from compas_timber.errors import FeatureApplicationError
 
 from .btlx_process import BTLxProcess
 from .btlx_process import BTLxProcessParams
@@ -441,7 +441,7 @@ class StepJointNotch(BTLxProcess):
 
         Raises
         ------
-        :class:`~compas_timber.elements.FeatureApplicationError`
+        :class:`~compas_timber.errors.FeatureApplicationError`
             If the cutting planes do not create a volume that itersects with beam geometry or any step fails.
 
         Returns
